Convert instant to discrete in numerous metrics
When using the pmlogger log-once directive for discrete metrics, it
helps alot if they actually are exported as discrete metrics.  When
they are not, the client tools are unable to operate correctly on
the daily archives produced by pmlogger_daily, as interpolation is
attempted (but cannot be done with just the one value).

This fixes a fistful of kernel metrics for Linux, Mac and Windows
that exhibited this problem.