Resolve path naming issues with more recent versions of autoconf.
Current versions of autoconf (2.61 for example) have added another
level of shell escaping and indirection to certain predefined paths
like ${prefix}.  One case is ${datadir}, which is now defined in
terms of ${datarootdir}, which is defined in terms of ${prefix}.

The sed filtering needed to be changed to correctly handle the man
page variable using ${mandir} - this is two levels deep now, and we
were producing the path NONE/share/man in pcp.conf before this.

This change is critical for current Debian/Ubuntu and MacOSX ports.