Rearrange source code so libpcp_pmc compiles on Mac OS X.
This change rearranges the libpcp_pmc header files to make the
build work on filesystems that are case-insensitive (or can be),
as is the default in Mac OS X.  The problem is String.h matches
the system header file string.h, both of which need to be pulled
into String.c++.  Including <string> doesnt work (tried that, it
looks like the compiler is too smart) either.  So, I've changed
the source structure to match the installed include structure,
and changed any #includes of libpcp_pmc headers to specify the
full <pcp/pmc/....h> header file path.