Allow pmie and/or operators to function with some data missing.
This change modifies pmie to allow a logical OR expression to evalute
to true when only once side of the expression tree can be evaluated,
due to host down / instance unavailable / insufficient samples.  Same
for logical AND expressions, and evaluating to false.

In order to test this modification, a new metric has been added into
the sample agent - sample.darkness - which shares an instance domain
with sample.color, but always returns no values available.