This is a fairly major rework of the way we manage contexts in kmchart.  It completely removes the need for source.{h,cpp} as we are now relying on all QmcContext data structure management being done within the metrics class.  Turns out the global list of contexts, and current context pointers, that we maintained in the Source class in kmchart were duplicating pointers and lists that the metrics class was already maintaining.  Output from diffstat on this change says it all:    18 files changed, 170 insertions(+), 345 deletions(-)
So, this removes a bunch of code in kmchart, removes unneeded
data structure indirection, and fixes a context management bug
in the code that deals with selecting the new current context
(the bug was evident when multiple hosts were added into the
sourceComboBox, and then a chart was created with the second
or third, or fourth - it would still use the initial host as
the source of metrics).

There was some code in source.cpp dealing with user interface
updates (combo boxes, metric trees) - this has been moved to
the source files dealing with all other aspects of those parts
of the interface (where they really belonged anyway).

There is one remaining issue with the new/edit chart dialog I
still need to resolve due to this change, but lots is done and
I need to do an initial checkpoint commit now so I can context
switch temporarily.