waitpid on dying agents at reconfig (qa/296)
qa/296 fails with daemon PMDAs hanging round after a reconfigure (ie
kill -HUP to pmcd). It appears that they need to be wait()ed upon.
Prior to the addition of the condition  || ap->ipcType == AGENT_DSO
to HarvestAgents in agent.c this condition was masked by other cleanup.
Previously, CleanupAgent did not waitpid() for AT_CONFIG. In addition,
there is no reason to sleep every time round the waitpid() loop, and in
many cases even once may be unnecessary