Fix the issue where, when started in archive mode, a Live tab's time axis wasn't being updated at all (i.e. without a live PCP context).
Also, made the changes necessary to make the archive time axis look
decent when started in Live mode (no archive sources).  This needed
a way to make the canvas hidden by default and sets the label on the
time axis to "[No open archives]".  Wasn't possible to just subclass
QwtPlot for this, with an updateLayout() that re-hid the canvas, cos
hide() on the canvas() resulted in a signal to updateLayout(), which
made us constantly call back to updateLayout() and a big jump in CPU
utilisation resulted.

As a fortunate by-product of the libqwt change, this fix also solves
the time axis rendering issue in the export image code (and print as
well, presumably).