When using a subsecond interval, we need a fuzzy-time-match tolerance that isn't larger than the delta.
Fine-tune the code that selects a matching time point in live
mode when changing "world view", matching up already-observed
values to required time points.

Finally, we only adjust chart state now when not entering the
stopped state.  This resolves the problem of getting an extra
(unintential) step in by pressing stop in the time controls.