Introduce an extended host name specification, allowing -h options that include pmcd port numbers, pmproxy host name, and proxy ports. See pmparsehostspec(3) for details of __pmParseHostSpec which does the actual parsing of hostspec to hosts and ports.  This provides for future possibility of multiple pmproxy hosts, but doesn't add the necessary pmproxyd extensions to support that.
I've implemented this as an impl.h interface (internal), as there
is no need to use it directly in clients (they just pass the new
format host string into pmNewContext), so I cunningly dodged the
need to increment libpcp's version number in the process.

Also, I changed pmFreeMetricSpec to not be a #define, and made it
a real routine (with type checking!), instead - it's still just a
wrapper to free(3) though.