Fix a problem where multiple pmie processes could be incorrectly launched for the same host, by pmie_check.sh.  A running pmie with a log filename matching the control file entry uniquely identifies that entry.  Adding default-hostname matching into this equation (as well) is unnecessary and turns out it can introduce a spurious problem when hostname aliases are being used.
In particular, if the pmhostname(1) call to gethostbyname(2) fails,
then the given hostname argument is printed out rather than the
name returned from gethostbyname.  This means pmhostname can return
different answers depending on the DNS state at the time, which in
turn can cause pmie_check to miss matching on a running pmie, and
incorrectly start a duplicate.

This was observed in our production environment and the fix has
been running there for a few days without the problem recurring
(previously it happened once or twice a day in some locations).