Add pmloglabel, a new tool which verifies, reports on, and can modify all details of the labels in each of the files of a Performance Co-Pilot (PCP) archive log.  The archive log has the base name archive and must have been previously created using pmlogger(1).
Each of the files in a PCP archive (metadata, temporal index, and one
or more data volumes) must contain a valid label at the start, else the
PCP tools will refuse to open the archive at all.

Thus, the primary function of pmloglabel is to be able to repair any
inconsistent or corrupt label fields, such that the entire archive is
not lost.  It will not check the remainder of the archive, but it will
give you a fighting chance to recover otherwise lost data.
Together, pmloglabel and pmlogextract are able to produce a valid PCP
archive from many forms of corruption.

 $ pmdumplog -l 20080125
 pmdumplog: Cannot open archive "20080125": Illegal label record at start of a PCP archive log file
 $ pmloglabel 20080125
 Mismatched PID (5264/5011) between temporal index and data volume 0
 $ pmloglabel -p 5264 20080125
 $ pmdumplog -l 20080125
 Log Label (Log Format Version 2)
 Performance metrics from host fw1
   commencing Fri Jan 25 00:10:09.341 2008
   ending     Sat Jan 26 00:09:54.344 2008