Rework the way the unified time axis is constructed.  We're now using a subclass of QStatusBar, and the time axis and button are children of this class.
This is a first step in a planned complete redesign of QLayout
management in kmchart, for which we need to be starting from a
more conventional QApplication layout (i.e. with a status bar
and with one real "central" widget, the tab widget).

The change has also had several beneficial side-effects, mainly
from just looking at this code once more:
- the time axis and date/label values now sit atop a toolbutton
widget, which means we can take action when pressed() and also
it usually inherits the nice shaded gradient look from the theme.
- resolved the time button size issue, such that the icon is no
longer scaled down by two pixels, resulting in a fuzzy font.
- fixed the value-label-clearing timeout, which regressed with
the kmsnap changes.