/etc/init.d/pmie status fails if there's more than one PMIE running.
The fix seems to be at line 372 of /usr/share/pcp/lib/rc-proc.sh.

<        [ $_n -eq 1 ] && return 0
>        [ $_n -ge 1 ] && return 0

RC_CHECKPROC claims to confirm to the LSB definition of rc_checkproc,
but LSB doesn't define an rc_checkproc or a checkproc.
http://refspecs.linux-foundation.org/LSB_3.2.0/LSB-Core-generic/LSB-Core-generic/iniscrptfunc.html

The closest thing there is is checkproc on SUSE, which says exit code 0
means "process is running" without making it explicit whether than means
one or many.
http://www.freebsd.org/cgi/man.cgi?query=checkproc&apropos=0&sektion=0&manpath=SuSE+Linux%2Fi386+8.2&format=html