Our production monitoring came across a situation where the filesystem instance domain had got thoroughly confused.  It looks like some (unknown) combination of mounts and unmounts (possibly at different places) had caused two different dev names to be pointing at the same entry in the table, and to be incorrectly reporting the same value.
The root cause isn't exactly known, however when looking at
the code and attempting to reproduce this, it became clear
that we're not using the pmdaCache interfaces here and we're
not correctly preserving the internal/external instance ID
mapping.

So, to fix this I've rewritten (much simplified) the code to
manage the filesys indom using the pmdaCache interfaces.
This obviously resolves the second problem, and very likely
the first as well.