Make the rpc/nfsd server statistics match up with what the kernel is exporting in current mainline Linux kernels (and still support older versions, of course).  This means some new rpc server metrics were added - ie. rpc.server.fh_anon, fh_nocache_dir and fh_nocache_nondir.  Additional metrics were also added for nfsd read/write bytes (total), as well as th_cnt (nfsd thread count) and th_fullcnt (number of times all threads were used).
There's yet more stuff in there, but for now this satisfies
all of my immediate problem solving needs.  The histogram
of nfsd thread utilisation might be useful to add, but its
a more invasive change (needs its own indom) and not on the
shopping list of things I must have ATM (might be useful to
the SGI storage guys though?).