Last major piece of surgery on the Perl API, should be just bug fixes from here on (hopefully!).  This adds two major pieces of functionality - the ability to generate an ascii namespace on- the-fly, and an interface for efficiently mapping a pmid back to a metric name for PMDAs to use.
The former will be used in the Install process to generate the
namespace (this allows all aspects of every metric to be set in
the add_metric() call - units, semantics, help text and finally
the name as well - esp. useful for DB agents where names may be
modified at Install time).

The latter proved useful in writing database PMDAs, where the
name of the metric is embedded in a column name or row entry -
this interface makes the fetch callback much simpler.