Remove use on kernel spinlock and quota headers (kernel headers are a moving target and generally unsuitable for direct inclusion in userspace tools).  Fortunately its a straightforward change to define the couple of structures we need here locally.
Several other code tweaks in the project quota code (for
SGI folks to review closely please):
- fix the 1st arg in call to Q_XQUOTASYNC quotactl
- delay doing the inode sync until we know we must, giving
plenty of opportunity to bail on XFS filesystems not using
project quota.
- check return values from malloc/calloc
- clear the acct/enfd flags early on, to ensure they are
not accidentally left.
- short-circuit the code in the case where project quota
is not enabled, no point scanning projects file then.
- remove the persistent cache loading on prjquota indom
(not clear how this helps when filesystem indom is random
in its numbering & always has been...)
- simplify the pmdacache interface use as a result.

One remaining question is why was /etc/projid not used in
place of /etc/projects, to get the symbolic project names
instead of numerics?