Provide an optional alternative to building using a symlink to pcp-version in the top level directory.  For Win32 (no symlink support to speak of) we now create a hardlink tree below this location, instead of the single symlink, and build there.  The Debian build is also changed to use this, since it also didn't like the symlink approach (dpkg-buildpackages does a `pwd` in that directory to intuit the package version, and that gives "pcp" when using a symlink).  The Debian build no longer has to be done by a nathans-special Makepkgs wrapper script, and now nicely deposits .deb files in build/deb.
But, mainly, this change is to give us a functional packaging
component to the build on Win32.

Also in here, is a change to handle all the autoconf/configure
generated files the same way - only removed on "make distclean
/ realclean" - previously this was half-and-half (some removed
on "make clean", some were not).  This implements the original
clear intention - tha "make clean" leaves these files, and a
"realclean" or "distclean" takes 'em out.

Finally, the mangle-src invocation is removed as its no longer
needed (AIUI) and is just complicating some of the make macros.