Fix the pmlogger_daily problem with daylight saving transitions.
Change the handling of time in the past to not use pmdate, and
algorithmic changes to avoid the problems associated with the
transitions at both ends of daylight savings periods.  Culling
and compressing are now based on the mtime of the archives.
Independent of when the script is run, today's archives are
NEVER candidates for merging.  Resurrected the lost Aztec mode
in which a superior script will merge all archives (not just
yesterday's archives) by default.  Added -o to support the old
Spanish mode where only yesterday's archives are merge candidates.