Really fix the Infiniband PMDA / RPM botch this time.
The last attempt overlooked the fact that pcp.spec.in
is not used by configure.in, rather its used by the
makefile in that subdir, so the manual translation in
there needs to be aware of the HAVE_IBDEV macro.  It
also means we must use values zero and one, undefined
and one doesn't cut it.

Also, PMDAs should not be conditionally added into the
SUBDIRS macro in src/pmdas/GNUmakefile, as the source
will only sometimes make it into the src tarball that
way - fixed that for IB while I was there.