Improve last change to pmie2col after Kens review.
delim="$OPTARG" (quoting helps prevent badness ... same is true for
wid, but less of an issue there given the likely values of wid).

In the AWK, use %s for delim in all places instead of %c, as -d "--"
is valid ... works the same for the degenerate case of a single char
delimiter.