Extend pmwtf so that it can used on different timespans in each log.
This is useful in the case where a utilisation spike occured, and
you wish to compare a time range with the spike to another range
close to that time (e.g. just before/after, same day).  The current
script assumes we're comparing the same timeframes on two different
days which is not the case here.