Initial conversion of Oracle PMDA to Perl.
This PMDA is incomplete.  It is based on the original Oracle
PMDA that Jonathan wrote at SGI, which was recently released
under the GPL (Pro*C version).  Remaining work would involve
testing it, and fleshing out the rest of the fetch and indom
callbacks.

All metric and instance domain definitions are included, and
the man page is here.  It is now possible for someone who has
a need to monitor Oracle, and just basic Perl / PCP knowledge
to pick this code up and finish it (and maintain it).  Thanks
to the use of Perl and DBI we are shielded so completely from
the underlying Oracle version (and implementation in general)
that this PMDA can now be shipped with the core PCP package.

Contrary to popular mythology, there is no reason this PMDA
will not be able to operate at comparable speeds to the old C
implementation.  In fact, from looking thru some of the SQL in
the original PMDA, there's potential for it to run faster than
before.