Cleanup of PCP QA in preparation for porting - remove cruft.
Nowadays we install pmclient by default, so don't make a local
copy in QA.  We also enable local context in pminfo, so no need
for local (non-portable) version of that source in QA.
Pull in the installed builddefs for building sources, instead
of encoding platform-specific bits in the src makefile (took
out existing platform-specific bits from Solaris and IRIX too).
Remove some of the source symlinks that were created, as thats
not portable to all platforms.