pmimport - a tool/API for creating PCP logs from external data.
This is a straight port of the code SGI recently made available.
It is now platform-independent though (not clear it even worked
on Linux before?).  I've added it to the tree, but decided not
to add it to the installed set of tools/APIs/docs yet, until we
have a real-world use in oss-land (in proprietary-land, we had
sar plugins for various flavours of IRIX and Unicos).