QA test 519 intermittent failure fixed.
This test intermittently fails when the two hosts chosen do not
have filesystems that are less than 50% full - dumb assumption,
and has nothing to do with what the test is actually trying to
achieve.  Changed this to more than 0% full, which makes for a
deterministic test.  Also made the syslog path search consistent
with the other syslog test (583).