Use killall directly in pmclusterd initscript
The PCP_KILLALL_PROG variable must have been an SGI-ism.
This really isn't the ideal way to do this, as it isn't
chroot-safe.  Since pmclusterd really only works on Altix
ICE right now we'll leave it as is.  It should really be
using a pidfile in /var/run and doing kill -TERM $pid.