Complete the porting of MMV PMDA to Windows.
This covers all of the path name platform-independence,
switches to the generic memory map/unmap routines now in
libpcp, uses the generic rename wrapper, __pmSetProgname
is now used.

It also changes the tempnam call.  Previously, this would
have allocated a temp file in TMPDIR and then renamed it
into the PCP pmns directory - if these are on different
filesystems, this would fail.  Its easily avoided though,
by creating the temp files alongside the other pmns files
(in the same directory) which is highly likely to be in
the same filesystem.