Complete the porting of MMV client library to Windows.
Quite a large refactoring of code here too.  Since we now
need to link with libpcp (for Win32 routines) we can make
use of pmGetConfig (Max cool with this, or maybe that was
the beer talking).

Split mmv_stats_init into a few static helper routines,
making the flow of code much easier to follow for first-
timers like me.  Use the platform-independent path type
interfaces, memory mapped file interfaces.  And finally,
picked 1 coding style and stuck with it, twas a bit of a
mix-and-match (style now consistent with rest of PCP).