Improvements to the pmie and pmlogger startup scripts.
In particular, we're now less verbose in the case where these
(check) scripts are auto-run from cron but pmie/pmlogger may
be chkconfig'd off on that host.  Thinking about moving toward
an /etc/cron.d/pcp file which is installed always (trialling
this on pmie in our production environment at the moment, but
thats not in this checkin, pending some observation).

Also reworked the pmlogger script snippet that deals with
the $PCP_TMP_DIR/pmlogger/primary link.  Previously we had
some platform-specific shell code, now it works the same on
all.  We also deal with the case where this file is a hard
link (as on Windows, where limited symlink support exists)
by testing for file existance and not just a symlink.