Extend libpcp_gui to include time control state management code.
This is the code from pmval, but made generic through the use of
callback functions for the points that a client tool needs to
take actions (step - fetching data, etc), mainly to simplify the
changes needed to console tools to incorporate support for the
pmtime(1) time control.

Finally, pmstat is modified to also be able to use time controls
as well (-g and -p options, same as pmval).  Fair bit of testing
done to ensure pmval still functions, some improvements made too
to the now-shared finite state automaton.