Extend the Windows pdhlist.c utility sources for Windows testing.
At some point between Windows Server 2003 and 2008 the regular
expression syntax for the PdhExpandCounterPath has mutated.  Still
figuring out the exact extent of this, but it prevents the Windows
PMDA metrics with instances from functioning at all currently, due
to the old regexs for those no longer being accepted.  This change
extends a diagnostic utility to better assist this exercise.