Fix the pmchart font problem under X11 -- FINALLY!
Reported by *many* people over the years now, particularly
prevalent on Fedora though.  After a recent Debian upgrade
at last I have a system which exhibits the problem, and an
hour or so of prodding and poking uncovered the problem.

It turns out that the use of a global QFont object for the
small font was the cause.  This meant the constructor was
being run before any other objects, including QApplication
and QMainWindow objects ... and somehow this confused all
manner of things downstream, once those later constructors
did get run.  Converting this global to a pointer and then
calling the constructor at a more appropriate point in the
main() routine gets fonts displayed correctly.  Yippeeeee!