Tweak the bash auto-completion - local context and pmprobe support.
Use local context mode when completing metric names in bash,
so that it works (and straight-away, not with a 5 sec timeout)
when pmcd is not running locally.

Also added pmprobe as a command for which we expand metric names.

Note that on Debian systems at least, in order for the system-wide
auto-completion to use /etc/bash_completion.d/ files, a few lines
in /etc/bash.bashrc need to be un-commented.