Use native Windows path separators when opening pmchart views.
This change makes (conditional, not available on older libpcp) use
of the __pmPathSeparator() libpcp interface to ensure paths opened
on Windows are in the appropriate style.  It also redoes code that
assumes $HOME is set and uses QDir::home() instead - this was able
to cause a SIGSEGV on startup on Windows.

Finally, it also reorders the view search path so that the pmchart
locations are scanned ahead of the kmchart compatability paths.