Remove gcc warnings.
Most of these changes are totally benign and not necessary because the
variables are _really_ initialized before being used ... but until
the gcc pixies are able to only warn in situations where there really
is a problem, we have little choice.

There are a couple of very very low probability bugs corrected here also.