Revert lex.l changes.
flex does not behave the same way on all platforms, and there does
not appear to be any way to safely remove the warnings and leave the
code compiling correctly on all platforms.

We're stuck with warnings like this ...
<stdout>:1446: warning: ‘yyunput’ defined but not used
<stdout>:1489: warning: ‘input’ defined but not used
from the build I'm afraid for at least some versions of flex.