Improve the instance domain handling logic in the Windows PMDA.
Fix a bug where we'd not set V_COLLECTED for any but the first
instance for some instance domains.  Move the logic which sets
the instance domain to empty initially to a more sensible spot
and rework the instance PDU handling code.

Fixed a small memory leak in re-establishing instance domains.
Allow for more CPUs in the per-processor indom (up to 9999).
Add a fast path into the fetch callback for metrics with indoms
which now first looks to see if the instance ID maps directly
into the value table, instead of always scanning from instance
ID zero to the end.  Esp. helps some large domains like threads
and processes, but also a small improvement for others.