Final format modifications for MMV PMDA, for instances.
This change reworks the way instance domains are handled
in the MMV PMDA, to be more consistent with the way it
deals with metrics.

mmv_stats_init() now takes extra arguments for instances
(indoms, nindoms).  The ondisk format has an additional
TOC type for indoms, separate to instances (paralleling
the metrics / values separation).  In addition to making
a more consistent format thats easier to write the Java
API for, we can also now specify indom help text and can
thus now support everything that a "regular" PMDA can.

This change also separates out the ondisk format structs
from the client interface (mmv_dev.h vs mmv_stats.h) so
that clients never need to directly access the on-disk
file.  Values *are* still updated "in-place", of course,
but this is done via access to a pmAtomValue instead of
by exposing implementation details at the API level.

Man pages are updated to document the API tweaks and the
ondisk format changes.  This is expected to be the final
format change we need in order to replace 2 of our mmap-
based production PMDAs (Aconex and Cymetry viewer) - its
likely some minor PMDA tweaks still remain to completely
achieve that goal, but the ondisk format is done.

As an interesting by-product of these changes, the main
API routine mmv_stats_init() no longer allocates memory
at all - everything is done in-place in the mmap'd file.