Layout the MMV strings section more like the other sections.
Paul Cowan pointed out that the way the strings section was
layed out ondisk was inconsistent with the other sections,
and for no good reason - wasted a little space too with all
the unneeded TOC sections.  Fixed up to be one section for
all strings and the TOC count field for strings now makes a
bit more sense.