Increase domain from 8-bits to 9-bits.
Cleaned up the use of the top bit as a flag for both a pmID and
a pmInDom ... now using a 1 bit flag field instead of the 2-bit
"pad" field.  Changes required for PMNS parsing in libpcp and
internal handling of help text.

Introduced PMDA_INTERFACE_4, but no code using this yet.

Encoding and logic to handle dynamic PMNS entries introduced and
some local support in parsing and reporting local PMNS entries
(syntax for PMID in PMNS is <domain>:*:*) ... this means domain
number 511 (DYNAMIC_PMID) is now special and reserved.