Do not create infiniband pmda config file by default
In order to simplify monitoring in the simple case where a
machine has a single Infiniband interface, the config file is
no longer created by default.  Instead, only local ports are
monitored.  This does solve the problem where if an infiniband
card is replaced in a machine that the administrator no longer
has to remove the /var/lib/pcp/pmdas/ib/config file in order
to restore infiniband monitoring.

For more advanced usage where static instance naming is required
or remote Infiniband ports must be monitored a config file can
still be created, but it is the responsibility of the administrator
to update the GUIDs if an HCA replacement is done.