Rearrange platform-specific handling in pmdaproc.sh.
Keep everything that does platform-specific stuff in one
place now, run early on, instead of all over the shop.

The code that handles determining whether the PCP::PMDA
Perl module is installed is also reworked to not fail in
the case where perl_opt is set, but user doesn't select
that option when Install'ing a PMDA.  This bit Mark with
the simple PMDA, which has both Perl and C variants.