Correct some regular expressions in Samba PMDA.
Of particular note, some of the heading lines in the
smbstatus --profile output have trailing whitespace.
Otherwise, this PMDA is now fully functional; and as
it used the stock Samba utilities, is not dependent
on the many flavours of Samba shmem format anymore.
At the cost of a slight overhead increase, of course,
but the original was unmaintainable.

Thanks to James Peach for advise on redesigning this.