Add missing dependencies and runlevels into PCP start scripts.
With dependency based boot sequencing, I discovered what I believe is
a bug in the init.d scripts of this package.  The list of runlevels to
stop in is incomplete.  The script should stop in runlevel 1 to make
sure the service is started again when switching from runlevel 1 to
2-5.  Also, the script should probably stop during halt (0) and reboot
(6).  The package should start in runlevel 4, just like in runlevels
2, 3 and 5.  I suspect the list of dependencies are incomplete.  The
scripts used to start after the syslog collector was operational, and
should probably continue to do so.  Also, the scripts need files in
/usr/ and should depend on $remote_fs.

If the scripts themselves need to start in any specific order, the
relationship should be listed explicitly in the LSB header too.  Make
sure the update-rc.d calls in the postinst reflect this ordering if
any such dependencies should exist.

This patch implement the proposed change.  Without it, the script seem
to fail to handle runlevel 0, 1, 4 and 6 properly, and might start too
early in the boot.